Adding standalone model definition for fastconformer preprocessor.
fast conformer requires melspec extraction and subsampling before
attention-conv stack. This is difficult to implement in vllm
(requires cache aware impl of stft and conv2d with strides),
so the plan is to run preprocessor outside of vllm.
This is a definition of the fastconformer preprocessor that is
streaming friendly. It subtly differs from NeMo model definition:
* center=False for stft
* no padding in conv layers
This is needed for smooth streaming inference. See example of
model usage on how to use streaming buffer and cuda graphs.
Forward pass for cuda graph (preparing single frame of fast conformer
input with batch size=1 in decoding stage) takes 0.1ms.

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>